+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Add serde::Deserialize for Wkt and Geometry.
+* Add helper function for deserializing from WKT format into geo_types::Geometry.
 
 ## 0.9.0